Docs maintenance (#21)
* Rename readthedocs.yml to .readthedocs.yaml

re: change in readthedocs config API

* Change to use new readthedocs build spec API

* Drop Sphinx source_suffix config

The deprecated list of strings appears to have been removed in
Sphinx=8. The default `source_suffix = {'.rst': 'restructuredtext'}` is
appropriate for this project.

* Drop Sphinx language config item

The default to English is appropriate for this project.

* Pin versions of sphinx and optional extensions

As recommended for reproducible builds on readthedocs:
https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
and to ensure that our docs build works with readthedocs changes to default
project dependencies:
https://blog.readthedocs.com/defaulting-latest-build-tools/

* Fix redirected & broken docs links found by linkcheck

* Update docs example output from Sphinx

Updated Sphinx build and linkcheck commands example outputs.